Update post controller to allow updating post caption
This commit updates the `updateCaptionOfPost` function in the post controller to allow users to update the caption of a post. It checks if the user is the owner of the post and then updates the caption accordingly. The updated post is returned in the response along with a success message.